requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
gunicorn==20.1.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.90.0
requests==2.31.0